DingQ_BE/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Utilities
requests==2.31.0

# ML dependencies (CPU-only PyTorch handled in Dockerfile)
numpy==2.1.3
torch==2.7.1
transformers==4.53.0
pillow==11.2.1
opencv-python==4.11.0.86
scikit-learn==1.7.0

# Environment
python-dotenv==1.0.0
